ref #37 refactor getRemoteInfo and restet token moved to BaseClient
